cuda: fixes problems and misconceptions regarding the NS(CudaArgument*) class hierachy in cuda submodule
